% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peregrine_file.R
\name{get_peregrine_file}
\alias{get_peregrine_file}
\title{Download Enhancer-Gene Link File from PANTHER Peregrine Database}
\usage{
get_peregrine_file(version = 19)
}
\arguments{
\item{version}{Integer (17, 18, or 19). Which PANTHER enhancer-gene link
version to download.}
}
\value{
The full path to the downloaded \code{.tsv} file stored in the
\code{BiocFileCache}.
}
\description{
Download Enhancer-Gene Link File from PANTHER Peregrine Database
}
\examples{
if (interactive()) {
    f <- get_peregrine_file(19)
    f
    readLines(f, n = 3)
}
}
