% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gp_links.R
\name{build_gp_links}
\alias{build_gp_links}
\title{Build Gene-Peak Network (Enhancer, Promoter, Closest)}
\usage{
build_gp_links(pk, gn, ver = 19, enh_file = NULL)
}
\arguments{
\item{pk}{Either:
(1) A character vector of genomic coordinates ("chr1:1000-1200"), or
(2) A data.frame containing peak coordinates with columns
"chr","start","end" or a "PeakRegion" column
("chr1:1000-1200" style).}

\item{gn}{Character vector of gene symbols
(HGNC official symbols, e.g. "TP53", "FMNL2").}

\item{ver}{Integer (17, 18, or 19) - PANTHER enhancer-gene link version.}

\item{enh_file}{Optional local path to enhancer-gene link file (.tsv).}
}
\value{
A data.frame with columns: Peak, Gene, Src.
Closest-gene mappings are based on distance from each peak to gene TSS.
}
\description{
Build Gene-Peak Network (Enhancer, Promoter, Closest)
}
\examples{
data("gp_example_inputs", package = "GPlinksR")
data("gp_example_links", package = "GPlinksR")
pk <- gp_example_inputs$pk[seq_len(3)]
gn <- gp_example_inputs$gn[seq_len(4)]

head(gp_example_links)

if (interactive()) {
    gp <- build_gp_links(pk, gn)
    head(gp)
}
}
