% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peregrine_file.R
\name{get_peregrine_file}
\alias{get_peregrine_file}
\title{Download Enhancer–Gene Link File from PANTHER Peregrine Database}
\usage{
get_peregrine_file(version = 19, destdir = tempdir())
}
\arguments{
\item{version}{Integer (17, 18, or 19). Which PANTHER enhancer–gene link version to download.}

\item{destdir}{Directory to save the file. Defaults to a temporary directory.}
}
\value{
The full path to the downloaded .tsv file.
}
\description{
Download Enhancer–Gene Link File from PANTHER Peregrine Database
}
\examples{
file_path <- get_peregrine_file(19)
head(readLines(file_path), 3)
}
